+++ src/Application.php
@@ -54,7 +54,7 @@ class RequestIdClass {
 class Application extends AbstractBaseClass implements ApplicationInterface
 {
   /** @const      string          Application/Framework version */
-  const VERSION = '0.0.18';
+  const VERSION = '0.0.19';
 
   /** @var        string          Application Environment (from ENV vars) */
   protected $environment;