+++ src/Helpers/UUID.php
@@ -1,54 +1,39 @@
 <?php declare(strict_types=1);
 
 /**
- * UUID
- *
- *   Generates a v4 UUID
- *
- * Exmaple:
- *   $uuidv4 = \\Spin\\Helper\\UUID::generate();                  // v4 UUID
- *   $uuidv5 = \\Spin\\Helper\\UUID::v5($uuidv4,'My v5 UUID');    // v5 UUID
+ * UUID Library
  *
  * @package  Spin
  */
 
 namespace Spin\Helpers;
 
+use \Ramsey\Uuid\Uuid AS RamseyUUID;
 use \Spin\Helpers\UUIDInterface;
 
 class UUID implements UUIDInterface
 {
   /**
-   * Generate v4 UUID
+   * Generate v6 UUID
    *
    * @return     string
    */
   public static function generate(): string
   {
-    # UUID generation based on random_bytes() method
-    $uuid = \vsprintf('%s%s-%s-%s-%s-%s%s%s',\str_split(\bin2hex(\random_bytes(16)), 4));
+    return self::v6();
+  }
 
-    return $uuid;
-
-    // return sprintf( '%04x%04x-%04x-%04x-%04x-%04x%04x%04x',
-    //     // 32 bits for "time_low"
-    //     \mt_rand( 0, 0xffff ), \mt_rand( 0, 0xffff ),
-
-    //     // 16 bits for "time_mid"
-    //     \mt_rand( 0, 0xffff ),
-
-    //     // 16 bits for "time_hi_and_version",
-    //     // four most significant bits holds version number 4
-    //     \mt_rand( 0, 0x0fff ) | 0x4000,
-
-    //     // 16 bits, 8 bits for "clk_seq_hi_res",
-    //     // 8 bits for "clk_seq_low",
-    //     // two most significant bits holds zero and one for variant DCE1.1
-    //     \mt_rand( 0, 0x3fff ) | 0x8000,
-
-    //     // 48 bits for "node"
-    //     \mt_rand( 0, 0xffff ), \mt_rand( 0, 0xffff ), \mt_rand( 0, 0xffff )
-    // );
+  /**
+   * Generate v3 UUID
+   *
+   * @param      string  $namespace  A Valid UUID
+   * @param      string  $name       A Random String
+   *
+   * @return     string
+   */
+  public static function v3(string $namespace, string $name): string
+  {
+    return RamseyUUID::uuid3($namespace,$name)->toString();
   }
 
   /**
@@ -58,7 +43,7 @@ class UUID implements UUIDInterface
    */
   public static function v4(): string
   {
-    return self::generate();
+    return RamseyUUID::uuid4()->toString();
   }
 
   /**
@@ -71,50 +56,28 @@ class UUID implements UUIDInterface
    */
   public static function v5(string $namespace, string $name): string
   {
-    if(!self::is_valid($namespace)) return false;
-
-    // Get hexadecimal components of namespace
-    $nhex = \str_replace(array('-','{','}'), '', $namespace);
-
-    // Binary Value
-    $nstr = '';
-
-    // Convert Namespace UUID to bits
-    for($i = 0; $i < \strlen($nhex); $i+=2) {
-      $nstr .= \chr(\hexdec($nhex[$i].$nhex[$i+1]));
-    }
-
-    // Calculate hash value
-    $hash = \sha1($nstr . $name);
-
-    return
-      \sprintf('%08s-%04s-%04x-%04x-%12s',
-      // 32 bits for "time_low"
-      \substr($hash, 0, 8),
-      // 16 bits for "time_mid"
-      \substr($hash, 8, 4),
-      // 16 bits for "time_hi_and_version",
-      // four most significant bits holds version number 5
-      (\hexdec(\substr($hash, 12, 4)) & 0x0fff) | 0x5000,
-      // 16 bits, 8 bits for "clk_seq_hi_res",
-      // 8 bits for "clk_seq_low",
-      // two most significant bits holds zero and one for variant DCE1.1
-      (\hexdec(\substr($hash, 16, 4)) & 0x3fff) | 0x8000,
-      // 48 bits for "node"
-      \substr($hash, 20, 12)
-    );
+    return RamseyUUID::uuid5($namespace,$name)->toString();
   }
 
   /**
-   * Checks if an UUID is valid (v3,v4 and v5)
+   * Generate a v6 UUID
    *
-   * @param      string  $uuid
+   * @return     string
+   */
+  public static function v6(): string
+  {
+    return RamseyUUID::uuid6()->toString();
+  }
+
+  /**
+   * Checks if an UUID is valid
    *
-   * @return     bool
+   * @param      string   $uuid
+   *
+   * @return     bool     True if valid
    */
   public static function is_valid(string $uuid): bool
   {
-    return \preg_match('/^\{?[0-9a-f]{8}\-?[0-9a-f]{4}\-?[0-9a-f]{4}\-?'.
-                       '[0-9a-f]{4}\-?[0-9a-f]{12}\}?$/i', $uuid) === 1;
+    return RamseyUUID::isValid($uuid);
   }
 }